@@ -4,7 +4,7 @@
   "description": "It's laundree day!",
   "main": "app.js",
   "engines": {
-    "node": "6"
+    "node": "7.8"
   },
   "dependencies": {
     "@google/maps": "^0.3.1",
@@ -14,30 +14,29 @@
     "babelify": "^7.2.0",
     "bcrypt": "^1.0.0",
     "blob-to-buffer": "^1.2.6",
-    "body-parser": "~1.16.0",
+    "body-parser": "^1.16.0",
     "browserify": "^14.0.0",
     "config": "^1.20.1",
     "connect-flash": "^0.1.1",
     "connect-redis": "^3.0.2",
-    "cookie-parser": "~1.4.3",
+    "cookie-parser": "^1.4.3",
     "cors": "^2.8.1",
     "debug": "^2.2.0",
     "detect-node": "^2.0.3",
     "email-templates": "^2.3.0",
-    "express": "~4.14.0",
+    "express": "^4.14.0",
     "express-session": "^1.13.0",
     "handlebars": "^4.0.5",
     "ical-generator": "^0.2.9",
-    "isparta": "^4.0.0",
     "jdenticon": "^1.3.2",
-    "laundree-sdk": "0.0.22",
+    "laundree-sdk": "0.0.23",
     "locale": "^0.1.0",
     "moment": "^2.16.0",
     "moment-timezone": "^0.5.7",
     "mongoose": "^4.6.8",
     "morgan": "^1.8.0",
     "node-sass-middleware": "^0.11.0",
-    "nodemailer": "^3.0.0",
+    "nodemailer": "^4.0.1",
     "nodemailer-stub-transport": "^1.0.0",
     "opbeat": "^4.0.0",
     "passport": "^0.3.2",
@@ -52,10 +51,11 @@
     "react-dom": "^15.3.0",
     "react-intl": "^2.1.2",
     "react-redux": "^5.0.0",
-    "react-router": "^3.0.2",
+    "react-router": "^4.0.0",
+    "react-router-dom": "^4.0.0",
     "redis": "^2.6.2",
     "redux": "^3.5.2",
-    "redux-actions": "^1.0.0",
+    "redux-actions": "^2.0.1",
     "socket.io": "^1.4.5",
     "socket.io-client": "^1.4.5",
     "superagent": "^3.3.1",
@@ -72,30 +72,32 @@
     "chai-as-promised": "^6.0.0",
     "chai-things": "^0.2.0",
     "codeclimate-test-reporter": "^0.4.0",
-    "eslint": "^3.3.1",
-    "eslint-config-standard": "^6.0.0",
-    "eslint-config-standard-jsx": "^3.0.0",
+    "eslint": "^3.19.0",
+    "eslint-config-standard": "^10.0.0",
+    "eslint-config-standard-jsx": "^4.0.0",
     "eslint-config-standard-react": "^4.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-node": "^4.2.2",
     "eslint-plugin-promise": "^3.0.0",
-    "eslint-plugin-react": "^6.1.2",
-    "eslint-plugin-standard": "^2.0.0",
-    "faker": "^3.1.0",
+    "eslint-plugin-react": "^6.10.3",
+    "eslint-plugin-standard": "^3.0.0",
+    "faker": "^4.1.0",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
-    "gulp-babel-istanbul": "^1.3.1",
     "gulp-coverage": "^0.3.38",
     "gulp-env": "^0.4.0",
     "gulp-eslint": "^3.0.1",
     "gulp-inject-modules": "^1.0.0",
     "gulp-mocha": "^4.0.1",
     "gulp-uglify": "^2.0.0",
+    "istanbul": "^0.4.5",
     "mocha": "^3.0.2",
     "nightwatch": "^0.9.8",
     "run-sequence": "^1.1.5",
     "sauce-connect-launcher": "^1.0.0",
     "selenium-standalone": "^6.0.1",
-    "sinon": "^1.17.6",
-    "sinon-chai": "^2.8.0",
+    "sinon": "^2.1.0",
+    "sinon-chai": "^2.9.0",
     "supertest": "^3.0.0",
     "supertest-as-promised": "^4.0.0",
     "vinyl-buffer": "^1.0.0",
