@@ -2,40 +2,41 @@
  * Created by budde on 05/06/16.
  */
 const React = require('react')
-const TopNav = require('./TopNav')
-const Footer = require('./Footer')
-const {DocumentTitle} = require('./intl')
+const {IntlProvider} = require('react-intl')
+const {Provider} = require('react-redux')
+const locales = require('./../../locales')
+const {Route, Switch} = require('react-router')
+const Auth = require('../containers/Auth')
+const BaseApp = require('../containers/BaseApp')
+const NativeApp = require('../containers/NativeApp')
+const TermsAndConditions = require('../views/TermsAndConditions')
+const Privacy = require('../views/Privacy')
+const StateCheckRedirectRoute = require('../containers/StateCheckRedirectRoute')
 
 class App extends React.Component {
   render () {
-    return <DocumentTitle title='document-title.base'>
-      <div className={this.props.user ? '' : 'footer'}>
-        <TopNav
-          config={this.props.config}
-          user={this.props.user}
-          location={this.props.location}
-          currentLaundry={this.props.currentLaundry}
-          laundries={this.props.laundries}/>
-        {this.props.children}
-        {this.props.user ? null : <Footer />}
-      </div>
-    </DocumentTitle>
+    return <IntlProvider locale={this.props.locale} messages={locales[this.props.locale].messages}>
+      <Provider store={this.props.store}>
+        <Switch>
+          <StateCheckRedirectRoute
+            redirectTo='/'
+            test={({currentUser}) => !currentUser}
+            path='/auth'
+            component={Auth} />
+          <Route path='/privacy' component={Privacy}/>
+          <Route path='/terms-and-conditions' component={TermsAndConditions}/>
+          <Route path='/native-app' component={NativeApp}/>
+          <Route path='/laundries/:laundryId' component={BaseApp} />
+          <Route path='/' component={BaseApp} />
+        </Switch>
+      </Provider>
+    </IntlProvider>
   }
 }
 
 App.propTypes = {
-  location: React.PropTypes.object,
-  config: React.PropTypes.shape({
-    locale: React.PropTypes.string.isRequired,
-    returningUser: React.PropTypes.bool.isRequired
-  }),
-  children: React.PropTypes.any,
-  currentLaundry: React.PropTypes.string,
-  laundries: React.PropTypes.object,
-  user: React.PropTypes.shape({
-    id: React.PropTypes.string,
-    photo: React.PropTypes.string
-  })
+  locale: React.PropTypes.oneOf(locales.supported),
+  store: React.PropTypes.object.isRequired
 }
 
 module.exports = App
