@@ -27,7 +27,7 @@ passport.use(new FacebookStrategy({
   clientID: config.get('facebook.appId'),
   clientSecret: config.get('facebook.appSecret'),
   callbackURL: config.get('facebook.callbackUrl'),
-  profileFields: ['id', 'first_name', 'last_name', 'middle_name', 'email', 'gender', 'displayName', 'link', 'picture']
+  profileFields: ['id', 'first_name', 'last_name', 'middle_name', 'email', 'gender', 'displayName', 'link', 'picture.type(large)']
 }, oauthCallback))
 
 passport.use(new GoogleStrategy({
