@@ -17,6 +17,7 @@ function setupSocket (server) {
 }
 
 function fetchLaundry (currentUserId, laundryId) {
+  debug('Fetching laundry', currentUserId, laundryId)
   return UserHandler
     .findFromId(currentUserId)
     .then(user => {
@@ -133,20 +134,28 @@ function setupFunctions (socket) {
   )
   defineLaundrySocketFunction(socket, 'listMachines',
     laundry => laundry.fetchMachines().then(actions.listMachines))
+
+  defineLaundrySocketFunction(
+    socket,
+    'fetchLaundry',
+    laundry => actions.listLaundries([laundry]))
 }
 
 function find (_Handler, options, fields = []) {
-  let filter = {}
+  debug('Finding with options', options)
+  const filter = {}
   if (options && options.q && fields.length) {
-    filter = {
-      $or: fields
-        .map(field => {
-          const obj = {}
-          obj[field] = new RegExp(options.q, 'i')
-          return obj
-        })
-    }
+    filter.$or = fields
+      .map(field => {
+        const obj = {}
+        obj[field] = new RegExp(options.q, 'i')
+        return obj
+      })
   }
+  if (options && !options.showDemo) {
+    filter.demo = {$ne: true}
+  }
+  debug('Using filter', filter)
   const opts = options ? {limit: options.limit, skip: options.skip, sort: {_id: 1}} : undefined
   return Promise.all([_Handler.find(filter, opts), _Handler.fetchCount(filter)])
 }
@@ -158,11 +167,6 @@ function setupAdminFunctions (socket) {
     'listLaundries',
     options => find(LaundryHandler, options, ['name'])
       .then(([laundries, size]) => [actions.listLaundries(laundries), actions.updateLaundryListSize(size)]))
-
-  defineLaundrySocketFunction(
-    socket,
-    'fetchLaundry',
-    laundry => actions.listLaundries([laundry]))
 
   defineSocketFunction(
     socket,
