@@ -5,7 +5,7 @@
 const connect = require('react-redux').connect
 const LeftNav = require('../views/LeftNav')
 
-const mapStateToProps = ({users, laundries, currentUser}, {params: {laundryId}}) => {
+const mapStateToProps = ({users, laundries, currentUser}, {match: {params: {laundryId}}}) => {
   return {
     laundries,
     currentLaundry: laundryId,
