@@ -4,8 +4,9 @@
 
 const connect = require('react-redux').connect
 const Timetable = require('../views/Timetable')
-
-const mapStateToProps = ({laundries, machines, bookings, currentUser, users}, {params: {laundryId}, location: {query: {offsetDate}}}) => {
+const queryString = require('querystring')
+const mapStateToProps = ({laundries, machines, bookings, currentUser, users}, {match: {params: {laundryId}}, location: {search}}) => {
+  const {offsetDate} = queryString.parse(search && search.substr(1))
   return {laundry: laundries[laundryId], machines, bookings, offsetDate, currentUser, users}
 }
 
