@@ -5,7 +5,7 @@
 const connect = require('react-redux').connect
 const Users = require('../views/Users')
 
-const mapStateToProps = ({laundries, users, invites, currentUser, config: {locale}}, {params: {laundryId}}) => {
+const mapStateToProps = ({laundries, users, invites, currentUser, config: {locale}}, {match: {params: {laundryId}}}) => {
   return {
     laundry: laundries[laundryId],
     locale,
