@@ -2,7 +2,7 @@
  * Created by budde on 11/06/16.
  */
 const React = require('react')
-const {Link} = require('react-router')
+const {Link} = require('react-router-dom')
 const {ValidationForm, ValidationElement} = require('./validation')
 const {ValueUpdater} = require('./helpers')
 const sdk = require('../../client/sdk')
@@ -10,7 +10,6 @@ const {DocumentTitle, Input, Submit, Label} = require('./intl')
 const {FormattedMessage} = require('react-intl')
 
 class Reset extends ValueUpdater {
-
   constructor (props) {
     super(props)
     this.submitHandler = (evt) => {
