@@ -7,7 +7,7 @@ const {FormattedDate, FormattedMessage} = require('react-intl')
 const string = require('../../utils/string')
 const moment = require('moment-timezone')
 const {DropDown, DropDownTitle, DropDownContent, DropDownCloser} = require('./dropdown')
-const {Link} = require('react-router')
+const {Link} = require('react-router-dom')
 
 const TimetableHeader = (props) => {
   const machines = props.laundry.machines
@@ -55,7 +55,6 @@ TimetableHeader.propTypes = {
 }
 
 class CalendarNavigationElement extends React.Component {
-
   constructor (props) {
     super(props)
     this.state = {current: this.firstDate}
@@ -165,7 +164,6 @@ CalendarNavigationElement.propTypes = {
 }
 
 class TimeTableHeaderNav extends React.Component {
-
   get firstDate () {
     return this.props.dates[0]
   }
