@@ -1,13 +1,14 @@
 const React = require('react')
-const {Link} = require('react-router')
+const {Link} = require('react-router-dom')
 const {ValidationElement, ValidationForm} = require('./validation')
 const {ValueUpdater} = require('./helpers')
 const sdk = require('../../client/sdk')
 const {DocumentTitle, Modal, Submit, Input, Label} = require('./intl')
 const {FormattedMessage} = require('react-intl')
+const Loader = require('./Loader')
+const NotFound = require('./NotFound')
 
 class UserNameForm extends ValueUpdater {
-
   constructor (props) {
     super(props)
     this.onSubmit = (event) => {
@@ -56,7 +57,6 @@ UserNameForm.propTypes = {
 }
 
 class UserPasswordForm extends ValueUpdater {
-
   constructor (props) {
     super(props)
     this.onSubmit = (event) => {
@@ -144,7 +144,6 @@ UserPasswordForm.propTypes = {
 }
 
 class DeleteUser extends React.Component {
-
   constructor (props) {
     super(props)
     this.state = {modalOpen: false}
@@ -176,14 +175,12 @@ class DeleteUser extends React.Component {
       </div>
     </div>
   }
-
 }
 
 DeleteUser.propTypes = {
   user: React.PropTypes.object.isRequired
 }
 class UserSettings extends React.Component {
-
   constructor (props) {
     super(props)
     this.state = {}
@@ -332,7 +329,6 @@ class UserSettings extends React.Component {
           </li>)}
     </ul>
   }
-
 }
 
 UserSettings.propTypes = {
@@ -342,4 +338,25 @@ UserSettings.propTypes = {
   users: React.PropTypes.object
 }
 
-module.exports = UserSettings
+const UserSettingsWrapper = ({user, currentUser, laundries, users}) => {
+  if (!users[user]) {
+    return <NotFound />
+  }
+  return <UserSettings user={user} users={users} laundries={laundries} currentUser={currentUser}/>
+}
+
+UserSettingsWrapper.propTypes = UserSettings.propTypes
+
+const UserSettingsLoaderWrapper = ({user, currentUser, laundries, users}) => {
+  const cUser = users[currentUser]
+  if (cUser.role !== 'admin') {
+    return <UserSettingsWrapper user={user} users={users} laundries={laundries} currentUser={currentUser}/>
+  }
+  return <Loader loader={() => sdk.fetchUser(user)}>
+    <UserSettingsWrapper user={user} users={users} laundries={laundries} currentUser={currentUser}/>
+  </Loader>
+}
+
+UserSettingsLoaderWrapper.propTypes = UserSettings.propTypes
+
+module.exports = UserSettingsLoaderWrapper
