@@ -3,21 +3,23 @@ const router = express.Router()
 const base64UrlSafe = require('urlsafe-base64')
 const {LaundryHandler} = require('../handlers')
 
+const notFoundError = new Error('Not found')
+notFoundError.status = 404
+
 router.get('/:laundryId/:id', (req, res, next) => {
   const {id, laundryId} = req.params
   if (!base64UrlSafe.validate(id) || !base64UrlSafe.validate(laundryId)) {
-    const error = new Error('Not found')
-    error.status = 404
-    return next(error)
+    return next(notFoundError)
   }
   if (!req.user) return res.redirect(`/auth?to=${encodeURIComponent(req.originalUrl)}`)
+  if (req.user.isDemo) {
+    return next(notFoundError)
+  }
   LaundryHandler
     .findFromShortId(laundryId)
     .then(laundry => {
       if (!laundry) {
-        const error = new Error('Not found')
-        error.status = 404
-        throw error
+        throw notFoundError
       }
       return laundry
         .verifyInviteCode(id)
