@@ -20,13 +20,19 @@ module.exports = {
         return u.generateVerifyEmailToken(email)
           .then(token => u.verifyEmail(email, token.secret))
           .then(() => u.createLaundry(faker.name.findName()))
-          .then(l => Promise
-            .all([
-              l.createMachine('M1', 'wash', true),
-              l.createMachine('M2', 'wash', false)
-            ]))
+          .then(l => l.createMachine('M1', 'wash', false))
       })
       .then(() => done(), err => console.log(err))
+  },
+  'Can mark machine broken': client => {
+    setup(client)
+      .waitForElementPresent('.machine_list li .repair svg', timeout)
+      .click('.machine_list li .repair svg')
+      .waitForElementPresent('.machine_list li .broken', timeout)
+      .waitForElementPresent('.machine_list li .repair svg', timeout)
+      .click('.machine_list li .repair svg')
+      .waitForElementPresent('.machine_list li .machineForm:not(.broken)', timeout)
+      .end()
   },
   'Can go to machines': client => {
     setup(client).end()
@@ -37,17 +43,9 @@ module.exports = {
       .waitForElementPresent('#LaundryMain .create_machine .machineForm input[type=text]', timeout)
       .setValue('#LaundryMain .create_machine .machineForm input[type=text]', name)
       .submitForm('#LaundryMain .create_machine .machineForm')
-      .waitForElementPresent('#LaundryMain .machine_list li:nth-of-type(3) ', timeout)
-    client.assert.valueContains('#LaundryMain .machine_list li:nth-of-type(3) input[type=text]', name)
+      .waitForElementPresent('#LaundryMain .machine_list li:nth-of-type(2) ', timeout)
+    client.assert.valueContains('#LaundryMain .machine_list li:nth-of-type(2) input[type=text]', name)
     client
-      .end()
-  },
-  'Can mark machine broken': client => {
-    setup(client)
-      .click('.machine_list li:last-of-type .repair svg')
-      .waitForElementPresent('.machine_list li:last-of-type .broken', timeout)
-      .click('.machine_list li:last-of-type .repair svg')
-      .waitForElementNotPresent('.machine_list li:last-of-type .broken', timeout)
       .end()
   }
 }
