@@ -2,14 +2,18 @@
  * Created by budde on 26/08/16.
  */
 
-const USER_NOT_VERIFIED = 'User not verified'
-const USER_NOT_FOUND = 'User not found'
-const INVALID_EMAIL_PASSWORD_COMBINATION = 'Invalid email/password combination'
-const USER_DOES_NOT_EXISTS = "User doesn't exists"
+const USER_NOT_VERIFIED = 'auth.error.not-verified'
+const USER_NOT_FOUND = 'auth.error.user-not-found'
+const INVALID_EMAIL_PASSWORD_COMBINATION = 'auth.error.invalid-email-password-combination'
+const USER_DOES_NOT_EXISTS = 'auth.error.user-does-not-exists'
+const INVALID_VERIFICATION_LINK = 'auth.error.invalid-verification-link'
+const EMAIL_VERIFIED = 'auth.success.email-verified'
 
 module.exports = {
+  EMAIL_VERIFIED,
   USER_NOT_VERIFIED,
   USER_NOT_FOUND,
   INVALID_EMAIL_PASSWORD_COMBINATION,
-  USER_DOES_NOT_EXISTS
+  USER_DOES_NOT_EXISTS,
+  INVALID_VERIFICATION_LINK
 }
